mcp_servers/opensearch-service/requirements.txt
# FastMCP / MCP
fastmcp>=2.14.0
mcp>=1.23.0

# OpenSearch
opensearch-py>=2.3.0

# HTTP / async
httpx>=0.28.0
httpx-sse>=0.4.0
nest-asyncio>=1.6.0

# Pydantic & settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Cloud.ru OpenAI-compatible SDK
openai>=1.0.0

# Tracing (OpenTelemetry)
opentelemetry-api>=1.0.0

# Numeric / ML
numpy>=1.24.0

# Utils
requests>=2.30.0
tqdm>=4.65.0
